Move Lynx mime-type last.
It seems some desktops don't use the matching for magic strings in the
file contents, but only rely on the globs for the filenames. It also
seems order matters for those. Moving the Lynx mime-type to the end
gives the regular .PRG mime-type precedence, which is much closer to
correct, as the vast majority of files named *.prg will actually be PRG
files and not Lynx archives.

Issue:	#3